+++ src/Exception/SegmentTranslationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Obullo\Router\Exception;
+
+use InvalidArgumentException;
+
+class SegmentTranslationException extends InvalidArgumentException
+{
+}