+++ tests/Resources/en/routing.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+	'{dummy}' => 'dummy',
+	'{slug}' => 'slug',
+	'{test}' => 'test',
+	'{hello}' => 'hello',
+	'{world}' => 'world',
+];