+++ tests/Resources/tr/routing.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+	'{dummy}' => 'aptal',
+	'{slug}' => 'slag',
+	'{test}' => 'test',
+	'{hello}' => 'merhaba',
+	'{world}' => 'dunya',
+];